backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# AI and ML libraries
google-generativeai==0.7.2

# RAG and Vector Database - Simplified stable versions
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.6
sentence-transformers==2.7.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==3.17.4

# Database and storage
aiofiles==23.2.1

# Environment and configuration
python-dotenv==1.0.0

# Core dependencies
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.25.2
typing-extensions==4.9.0

# Additional utilities
numpy==1.24.4
requests==2.31.0

# Additional required dependencies
setuptools==68.0.0
wheel==0.41.2